pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cb-organizer"
version = "0.1.0"
description = "Local-first insurance and medical expense organizer"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.2",
  "jinja2>=3.1.4",
  "python-multipart>=0.0.9",
  "argon2-cffi>=23.1.0",
  "cryptography>=42.0.8",
  "keyring>=25.2.1",
  "apscheduler>=3.10.4",
  "pydantic>=2.8.2",
  "httpx>=0.27.0",
  "itsdangerous>=2.2.0",
  "pyotp>=2.9.0",
  "reportlab>=4.2.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "playwright>=1.45.0",
  "ruff>=0.5.7",
  "mypy>=1.10.1",
  "bandit>=1.7.9",
  "pyinstaller>=6.10.0",
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
markers = [
  "optional: optional smoke-style tests",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
strict_optional = true

[tool.bandit]
exclude_dirs = ["tests"]
